Various SidPoint based fixes
Improve display of sidpoint related info
Ensure total.cat files are generated correctly
Ensure catsid.py prefers more recent previous points

+++ catsid.py
@@ -7,21 +7,22 @@ from process import join
 from django.core.management import setup_environ
 import clouds.settings
 setup_environ(clouds.settings)
-from clouds.models import SidPoint, Line, SidTime 
+from clouds.models import RealPoint, SidPoint, Line, SidTime 
 
 from catlib import parse_cat
 
-from django.db import connection, transaction
+from django.db import connection, transaction, reset_queries
 cursor = connection.cursor()
-#cursor.execute("DELETE FROM clouds_realpoint")
-cursor.execute("DELETE FROM clouds_sidpoint")
-cursor.execute("DELETE FROM clouds_line")
+cursor.execute("DROP TABLE clouds_realpoint")
+cursor.execute("DROP TABLE clouds_sidpoint")
+cursor.execute("DROP TABLE clouds_line")
 transaction.commit_unless_managed()
+subprocess.call(['python', 'manage.py', 'syncdb', '--all'])
 
 outdir = 'out'
 
 def catsid(sidtime, prev_points_list, rerun=False):
-    print sidtime
+    print 'Sidtime', sidtime
     siddir = os.path.join(outdir, 'cat', sidtime.get_dir())
     
     path = os.path.join(siddir, 'total.cat')
@@ -39,28 +40,34 @@ def catsid(sidtime, prev_points_list, rerun=False):
             if sidpoint.prev:
                 continue
         matched = False
+        step = 0
         for prev_sidtime, prev_points in prev_points_list:
+            step += 1
             dsquared = (prev_points-point)**2
             distances = dsquared['x'] + dsquared['y']
+            # Maybe fall back to second closest etc.
             distance = distances.min()
             if distance < 3**2:
                 idx = distances.idxmin()
                 # This name is confusing
                 try:
                     prev_point = SidPoint.objects.get(sidtime=prev_sidtime, idx=idx)
-                except SidPoint.DoesNotExist, e: raise e #continue
-                if prev_point.next: # then we have a collision
-                    distance1 = ( (prev_point.x - prev_point.next.x)**2 + 
-                            (prev_point.y - prev_point.next.y)**2 )
-                    if distance < distance1:
-                        discard_point = prev_point.next 
-                        line = Line()
-                        line.save()
-                        discard_point.line = line
-                        discard_point.prev = None
-                        discard_point.save()
-                        line = prev_point.line
-                        matched = True
+                except SidPoint.DoesNotExist, e:
+                    #break
+                    raise e
+                competing_point = prev_point.next 
+                if competing_point: # then we have a collision
+                    if competing_point.step >= i:
+                        distance1 = ( (prev_point.x - competing_point.x)**2 + 
+                                (prev_point.y - competing_point.y)**2 )
+                        if distance < distance1:
+                            line = Line()
+                            line.save()
+                            competing_point.line = line
+                            competing_point.prev = None
+                            competing_point.save()
+                            line = prev_point.line
+                            matched = True
                 else:
                     line = prev_point.line
                     matched = True
@@ -69,31 +76,45 @@ def catsid(sidtime, prev_points_list, rerun=False):
             if rerun:
                 continue
             prev_point = None
+            step = None
             line = Line()
             line.save()
         if rerun:
             sidpoint.prev = prev_point
-            sidpoint.line.sidpoint_set.update(line=line)
-            sidpoint.line.delete()
-            sidpoint.line = line
+            if sidpoint.line != line:
+                sidpoint.line.sidpoint_set.update(line=line)
+                sidpoint.line.delete()
+                sidpoint.line = line
         else:
-            sidpoint = SidPoint(x=point['x'], y=point['y'], flux=point['flux'], line=line, idx=i, sidtime=sidtime, prev=prev_point) 
+            sidpoint = SidPoint(
+                x=point['x'], y=point['y'],
+                x_min=point['x_min'], y_min=point['y_min'],
+                width=point['width'], height=point['height'],
+                flux=point['flux'],
+                line=line, idx=i, sidtime=sidtime, prev=prev_point, step=step) 
         sidpoint.save()
-    #    if len(points_django) > 1000: # Avoid sqlite variable limit #SidPoint.objects.bulk_create(points_django)
-    #points_django = []
+
+    reset_queries()
 
     if not rerun:
         prev_points_list.insert(0, (sidtime, points))
-        if len(prev_points_list) > 4:
-            prev_points_list.pop()
+    if rerun or len(prev_points_list) > 4:
+        prev_points_list.pop()
 
 sidtimes = list(SidTime.objects.order_by('time'))
 prev_points_list = []
+"""
+for sidtime in sidtimes[:4]:
+    catsid(sidtime, prev_points_list)
+prev_points_list = []
+for sidtime in sidtimes[-5:]:
+    catsid(sidtime, prev_points_list)
+"""
 for sidtime in sidtimes:
     catsid(sidtime, prev_points_list)
 for sidtime in sidtimes[:4]:
     catsid(sidtime, prev_points_list, True)
+print prev_points_list
 
 
-#SidPoint.objects.bulk_create(points_django)
